+++ ingestor-default/src/main/scala/za/co/absa/hyperdrive/ingestor/implementation/transformer/avro/confluent/AdvancedAvroToSparkConverter.scala
@@ -14,28 +14,22 @@
  * limitations under the License.
  */
 
-package za.co.absa.hyperdrive.ingestor.implementation.utils
+package za.co.absa.hyperdrive.ingestor.implementation.transformer.avro.confluent
 
 import org.apache.avro.LogicalTypes.{Date, Decimal, TimestampMicros, TimestampMillis}
-import org.apache.avro.Schema.Field
+import org.apache.avro.Schema
 import org.apache.avro.Schema.Type._
-import org.apache.avro.SchemaBuilder.{FieldAssembler, FieldBuilder}
 import org.apache.avro.util.internal.JacksonUtils
-import org.apache.avro.{JsonProperties, LogicalTypes, Schema, SchemaBuilder}
-import org.apache.spark.sql.types.Decimal.minBytesForPrecision
 import org.apache.spark.sql.types._
-import org.codehaus.jackson.map.{ObjectMapper, ObjectReader}
-
-import java.io.{BufferedOutputStream, ByteArrayOutputStream}
-import java.util.Objects
+import org.codehaus.jackson.map.ObjectMapper
+import za.co.absa.abris.avro.sql.SchemaConverter
+import java.io.ByteArrayOutputStream
 import scala.collection.JavaConverters._
-import scala.util.Try
+import za.co.absa.hyperdrive.ingestor.implementation.transformer.avro.confluent.SparkMetadataKeys._
 
 // scalastyle:off
-object CustomSchemaConverters {
-  private lazy val nullSchema = Schema.create(Schema.Type.NULL)
-  private val primitiveType = "primitiveType"
-  private val defaultValueKey = "default"
+class AdvancedAvroToSparkConverter extends SchemaConverter {
+  override val shortName: String = AdvancedAvroToSparkConverter.name
   private lazy val objectMapper = new ObjectMapper()
 
   case class SchemaType(dataType: DataType, nullable: Boolean, primitiveType: Option[Schema])
@@ -43,7 +37,7 @@ object CustomSchemaConverters {
   /**
    * This function takes an avro schema and returns a sql schema.
    */
-  def toSqlType(avroSchema: Schema): DataType = {
+  override def toSqlType(avroSchema: Schema): DataType = {
     toSqlTypeHelper(avroSchema, Set.empty).dataType
   }
 
@@ -86,7 +80,7 @@ object CustomSchemaConverters {
             case Some(defaultJson) =>
               val baos = new ByteArrayOutputStream()
               objectMapper.writeValue(baos, defaultJson)
-              val r = metadataBuilder.putString(defaultValueKey, baos.toString)
+              val r = metadataBuilder.putString(DefaultValueKey, baos.toString)
               baos.close()
               r
             case None => metadataBuilder
@@ -95,7 +89,7 @@ object CustomSchemaConverters {
           val schemaType = toSqlTypeHelper(f.schema(), newRecordNames)
           schemaType.primitiveType
             .map(_.toString)
-            .map(schema => metadataBuilderWithDefault.putString(primitiveType, schema).build())
+            .map(schema => metadataBuilderWithDefault.putString(PrimitiveTypeKey, schema).build())
             .map(metadata => StructField(f.name, schemaType.dataType, schemaType.nullable, metadata))
             .getOrElse(StructField(f.name, schemaType.dataType, schemaType.nullable, metadataBuilderWithDefault.build()))
         }
@@ -140,7 +134,7 @@ object CustomSchemaConverters {
                 val schemaType = toSqlTypeHelper(s, existingRecordNames)
                 schemaType.primitiveType
                   .map(_.toString)
-                  .map(schema => new MetadataBuilder().putString(primitiveType, schema).build())
+                  .map(schema => new MetadataBuilder().putString(PrimitiveTypeKey, schema).build())
                   .map(metadata => StructField(s"member$i", schemaType.dataType, schemaType.nullable, metadata))
                    // All fields are nullable because only one of them is set at a time
                   .getOrElse(StructField(s"member$i", schemaType.dataType, nullable = true))
@@ -152,107 +146,9 @@ object CustomSchemaConverters {
       case other => throw new IncompatibleSchemaException(s"Unsupported type $other")
     }
   }
-
-  def toAvroType(
-    catalystType: DataType,
-    avroSchema: Option[Schema],
-    nullable: Boolean = false,
-    defaultValue: Option[Object] = None,
-    recordName: String = "topLevelRecord",
-    nameSpace: String = "")
-  : Schema = {
-    val builder = SchemaBuilder.builder()
-
-    val schema = catalystType match {
-      case BooleanType => builder.booleanType()
-      case ByteType | ShortType | IntegerType => builder.intType()
-      case LongType => builder.longType()
-      case DateType =>
-        LogicalTypes.date().addToSchema(builder.intType())
-      case TimestampType => avroSchema match {
-        case Some(schema) if schema.getLogicalType.isInstanceOf[TimestampMillis] =>
-          LogicalTypes.timestampMillis().addToSchema(builder.longType())
-        case _ => LogicalTypes.timestampMicros().addToSchema(builder.longType())
-      }
-      case FloatType => builder.floatType()
-      case DoubleType => builder.doubleType()
-      case StringType => builder.stringType()
-      case d: DecimalType => avroSchema match {
-        case Some(schema) if schema.getType == BYTES =>
-          val avroType = LogicalTypes.decimal(d.precision, d.scale)
-          avroType.addToSchema(SchemaBuilder.builder().bytesType())
-        case _ => getDecimalFixedType(d, avroSchema, nameSpace, recordName)
-      }
-      case BinaryType => avroSchema match {
-        case Some(schema) if schema.getType == FIXED =>
-          // Need to avoid naming conflict for the fixed fields
-          val name = nameSpace match {
-            case "" => s"$recordName.fixed"
-            case _ => s"$nameSpace.$recordName.fixed"
-          }
-          builder
-            .fixed(name)
-            .size(schema.getFixedSize)
-        case _ => builder.bytesType()
-      }
-      case ArrayType(et, containsNull) =>
-        builder.array()
-          .items(toAvroType(et, avroSchema, containsNull, None, recordName, nameSpace))
-      case MapType(StringType, vt, valueContainsNull) =>
-        builder.map()
-          .values(toAvroType(vt, avroSchema, valueContainsNull, None, recordName, nameSpace))
-      case st: StructType =>
-        val childNameSpace = if (nameSpace != "") s"$nameSpace.$recordName" else recordName
-        val fieldsAssembler = builder.record(recordName).namespace(nameSpace).fields()
-        st.foreach { f =>
-          val schema = Try(f.metadata.getString(primitiveType)).toOption
-            .map(schema => new Schema.Parser().parse(schema))
-          val defaultValueOpt = Try(f.metadata.getString(defaultValueKey)).flatMap(defaultJsonString => Try {
-            val jsonNode = objectMapper.readTree(defaultJsonString)
-            JacksonUtils.toObject(jsonNode)
-          }).toOption
-          val fieldAvroType =
-            toAvroType(f.dataType, schema, f.nullable, defaultValueOpt, f.name, childNameSpace)
-           defaultValueOpt match {
-            case Some(defaultObject) if !Objects.equals(defaultObject, JsonProperties.NULL_VALUE) =>
-              fieldsAssembler.name(f.name).`type`(fieldAvroType).withDefault(defaultObject)
-            case Some(_) =>
-              fieldsAssembler.name(f.name).`type`(fieldAvroType).withDefault(null)
-            case _ => fieldsAssembler.name(f.name).`type`(fieldAvroType).noDefault()
-          }
-
-        }
-        fieldsAssembler.endRecord()
-
-      // This should never happen.
-      case other => throw new IncompatibleSchemaException(s"Unexpected type $other.")
-    }
-    if (nullable) {
-      defaultValue match {
-        case Some(value) if !value.isInstanceOf[JsonProperties.Null] => Schema.createUnion(schema, nullSchema)
-        case _ => Schema.createUnion(nullSchema, schema)
-      }
-
-    } else {
-      schema
-    }
-  }
-
-  private def getDecimalFixedType(d: DecimalType, avroSchema: Option[Schema], nameSpace: String, recordName: String) = {
-    val avroType = LogicalTypes.decimal(d.precision, d.scale)
-    avroSchema.map(schema =>
-      avroType.addToSchema(SchemaBuilder.fixed(schema.getName).namespace(schema.getNamespace).size(schema.getFixedSize))
-    ).getOrElse {
-      val fixedSize = minBytesForPrecision(d.precision)
-      // Need to avoid naming conflict for the fixed fields
-      val name = nameSpace match {
-        case "" => s"$recordName.fixed"
-        case _ => s"$nameSpace.$recordName.fixed"
-      }
-      avroType.addToSchema(SchemaBuilder.fixed(name).size(fixedSize))
-    }
-  }
 }
 
-class IncompatibleSchemaException(msg: String, ex: Throwable = null) extends Exception(msg, ex)
-
+// scalastyle:on
+object AdvancedAvroToSparkConverter {
+  val name = "advanced"
+}
